Move most code related to legacy templating system into separete module.
All standard environments are augmented in the legacy_templating module.
Many functions related to the legacy templating system are decorated to
trigger deprecation warnings and to inform the developer about their
deprecated status.
