+++ flow/environment.py
@@ -12,14 +12,10 @@ environment, e.g. for the adjustemt of scheduler submission scripts.
 from __future__ import print_function
 from __future__ import division
 import os
-import sys
 import re
 import socket
 import logging
-import warnings
-import io
 import importlib
-from math import ceil
 from collections import OrderedDict
 
 from signac.common import config
@@ -33,9 +29,8 @@ from .scheduling.torque import TorqueScheduler
 from .scheduling.simple_scheduler import SimpleScheduler
 from .scheduling.fakescheduler import FakeScheduler
 from .util import config as flow_config
-from .util.template_filters import format_timedelta
-from .errors import SubmitError
 from .errors import NoSchedulerError
+from . import legacy_templating
 
 if six.PY2:
     import imp
@@ -49,22 +44,6 @@ if six.PY2:
 
 # Global variable can be used to override detected environment
 ENVIRONMENT = None
-
-
-NUM_NODES_WARNING = """Unable to determine the reqired number of nodes (nn) for this submission.
-Either provide this value directly with '--nn' or provide the number of processors
-per node: '--ppn'.
-
-Please note, you can ignore this message by specifying extra submission options
-with '--' or by using the '--force' option."""
-
-UTILIZATION_WARNING = """You either specified or the environment is configured to use {ppn}
-processors per node (ppn), however you only use {usage:0.2%} of each node.
-Consider to increase the number of processors per operation (--np)
-or adjust the processors per node (--ppn).
-
-Alternatively, you can also use --force to ignore this warning.
-"""
 
 
 def setup(py_modules, **attrs):
@@ -118,53 +97,8 @@ class ComputeEnvironmentType(type):
         return super(ComputeEnvironmentType, cls).__init__(name, bases, dct)
 
 
-class JobScript(io.StringIO):
-    """"Simple StringIO wrapper for the creation of job submission scripts.
-
-    Using this class to write a job submission script allows us to use
-    environment specific expressions, for example for MPI commands.
-    """
-    eol = '\n'
-
-    def __init__(self, env):
-        self._env = env
-        super(JobScript, self).__init__()
-
-    def __str__(self):
-        self.seek(0)
-        return self.read()
-
-    def write(self, s):
-        if six.PY2:
-            super(JobScript, self).write(unicode(s))  # noqa
-        else:
-            super(JobScript, self).write(s)
-
-    def writeline(self, line=''):
-        "Write one line to the job script."
-        self.write(line + self.eol)
-
-    def write_cmd(self, cmd, bg=False, np=None):
-        """Write a command to the jobscript.
-
-        This command wrapper function is a convenience function, which
-        adds mpi and other directives whenever necessary.
-
-        :param cmd: The command to write to the jobscript.
-        :type cmd: str
-        :param np: The number of processors required for execution.
-        :type np: int
-        """
-        if np is not None:
-            warnings.warn(DeprecationWarning("Do not provide np with write_cmd()!"))
-            if np > 1:
-                cmd = self._env.mpi_cmd(cmd, np=np)
-        if bg:
-            cmd = self._env.bg(cmd)
-        self.writeline(cmd)
-
-
-class ComputeEnvironment(with_metaclass(ComputeEnvironmentType)):
+class ComputeEnvironment(with_metaclass(ComputeEnvironmentType,
+                                        legacy_templating.LegacyComputeEnvironment)):
     """Define computational environments.
 
     The ComputeEnvironment class allows us to automatically determine
@@ -180,15 +114,6 @@ class ComputeEnvironment(with_metaclass(ComputeEnvironmentType)):
     hostname_pattern = None
     submit_flags = None
     template = 'base_script.sh'
-
-    @classmethod
-    def script(cls, **kwargs):
-        """Return a JobScript instance.
-
-        Derived ComputeEnvironment classes may require additional
-        arguments for the creation of a job submission script.
-        """
-        return JobScript(cls)
 
     @classmethod
     def is_present(cls):
@@ -221,6 +146,7 @@ class ComputeEnvironment(with_metaclass(ComputeEnvironmentType)):
                 "No scheduler defined for environment '{}'.".format(cls.__name__))
 
     @classmethod
+    @legacy_templating.support_legacy_templating_submit
     def submit(cls, script, flags=None, *args, **kwargs):
         """Submit a job submission script to the environment's scheduler.
 
@@ -234,16 +160,8 @@ class ComputeEnvironment(with_metaclass(ComputeEnvironmentType)):
             flags.extend(env_flags)
 
         # Hand off the actual submission to the scheduler
-        if isinstance(script, JobScript):  # api version < 6
-            script = str(script)
-
         if cls.get_scheduler().submit(script, flags=flags, *args, **kwargs):
             return JobStatus.submitted
-
-    @staticmethod
-    def bg(cmd):
-        "Wrap a command (cmd) to be executed in the background."
-        return cmd + ' &'
 
     @classmethod
     def add_args(cls, parser):
@@ -300,12 +218,12 @@ class UnknownEnvironment(StandardEnvironment):
     "Deprecated 'standard' environment, replaced by 'StandardEnvironment.'"
 
     def __init__(self, *args, **kwargs):
-        super(
-            "The 'UnknownEnvironment' class has been replaced by the "
-            "'StandardEnvironment' class.", DeprecationWarning)
-        super(UnknownEnvironment, self).__init__(*args, **kwargs)
+        raise RuntimeError(
+            "The 'flow.environment.UnknownEnvironment' class has been replaced by the "
+            "'flow.environment.StandardEnvironment' class.")
 
 
+@legacy_templating.support_legacy_templating
 class TestEnvironment(ComputeEnvironment):
     """This is a test environment.
 
@@ -315,17 +233,6 @@ class TestEnvironment(ComputeEnvironment):
     a real scheduler.
     """
     scheduler_type = FakeScheduler
-
-    @classmethod
-    def mpi_cmd(cls, cmd, np):
-        return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
-
-    @classmethod
-    def script(cls, **kwargs):
-        js = super(TestEnvironment, cls).script(**kwargs)
-        for key in sorted(kwargs):
-            js.writeline('#TEST {}={}'.format(key, kwargs[key]))
-        return js
 
 
 class SimpleSchedulerEnvironment(ComputeEnvironment):
@@ -364,46 +271,17 @@ class LSFEnvironment(ComputeEnvironment):
     template = 'lsf.sh'
 
 
+@legacy_templating.support_legacy_templating
 class NodesEnvironment(ComputeEnvironment):
     """A compute environment consisting of multiple compute nodes.
 
     Each compute node is assumed to have a specific number of compute units, e.g., CPUs.
     """
 
-    @classmethod
-    def calc_num_nodes(cls, np_total, ppn, force=False, **kwargs):
-        if ppn is None:
-            try:
-                ppn = getattr(cls, 'cores_per_node')
-            except AttributeError:
-                raise SubmitError(NUM_NODES_WARNING)
 
-        # Calculate the total number of required nodes
-        nn = int(ceil(np_total / ppn))
-
-        if not force:  # Perform basic check concerning the node utilization.
-            usage = np_total / nn / ppn
-            if usage < 0.9:
-                print(UTILIZATION_WARNING.format(ppn=ppn, usage=usage), file=sys.stderr)
-                raise SubmitError("Bad node utilization!")
-        return nn
-
-    @classmethod
-    def add_args(cls, parser):
-        super(NodesEnvironment, cls).add_args(parser)
-        parser.add_argument(
-            '--nn',
-            type=int,
-            help="Specify the total number of nodes. This value is computed automatically "
-                 "from the 'np' directive otherwise.")
-
-
+@legacy_templating.support_legacy_templating
 class DefaultTorqueEnvironment(NodesEnvironment, TorqueEnvironment):
     "A default environment for environments with TORQUE scheduler."
-
-    @classmethod
-    def mpi_cmd(cls, cmd, np):
-        return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
 
     @classmethod
     def add_args(cls, parser):
@@ -426,46 +304,10 @@ class DefaultTorqueEnvironment(NodesEnvironment, TorqueEnvironment):
             action='store_true',
             help="Do not copy current environment variables into compute node environment.")
 
-    @classmethod
-    def gen_tasks(cls, js, np_total):
-        """Helper function to generate the number of tasks (for overriding)"""
-        js.writeline('#PBS -l nodes={}'.format(np_total))
-        return js
 
-    @classmethod
-    def script(cls, _id, np_total, walltime=None, no_copy_env=False, **kwargs):
-        js = super(DefaultTorqueEnvironment, cls).script()
-        js.writeline('#PBS -N {}'.format(_id))
-        js = cls.gen_tasks(js)
-        if walltime is not None:
-            js.writeline('#PBS -l walltime={}'.format(format_timedelta(walltime)))
-        if not no_copy_env:
-            js.writeline('#PBS -V')
-        return js
-
-
+@legacy_templating.support_legacy_templating
 class DefaultSlurmEnvironment(NodesEnvironment, SlurmEnvironment):
     "A default environment for environments with SLURM scheduler."
-
-    @classmethod
-    def mpi_cmd(cls, cmd, np):
-        return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
-
-    @classmethod
-    def gen_tasks(cls, js, np_total):
-        """Helper function to generate the number of tasks (for overriding)"""
-        js.writeline('#SBATCH --ntasks={}'.format(np_total))
-        return js
-
-    @classmethod
-    def script(cls, _id, np_total, walltime=None, **kwargs):
-        js = super(DefaultSlurmEnvironment, cls).script()
-        js.writeline('#!/bin/bash')
-        js.writeline('#SBATCH --job-name="{}"'.format(_id))
-        js = cls.gen_tasks(js, np_total)
-        if walltime is not None:
-            js.writeline('#SBATCH -t {}'.format(format_timedelta(walltime)))
-        return js
 
     @classmethod
     def add_args(cls, parser):
@@ -486,23 +328,9 @@ class DefaultSlurmEnvironment(NodesEnvironment, SlurmEnvironment):
                  "completion of a cluster job with this id.")
 
 
+@legacy_templating.support_legacy_templating
 class DefaultLSFEnvironment(NodesEnvironment, LSFEnvironment):
     "A default environment for environments with LSF scheduler."
-
-    @classmethod
-    def mpi_cmd(cls, cmd, np):
-        raise NotImplementedError("LSF environments are not supported by the "
-                                  "legacy templating system.")
-
-    @classmethod
-    def gen_tasks(cls, js, np_total):
-        raise NotImplementedError("LSF environments are not supported by the "
-                                  "legacy templating system.")
-
-    @classmethod
-    def script(cls, _id, np_total, walltime=None, **kwargs):
-        raise NotImplementedError("LSF environments are not supported by the "
-                                  "legacy templating system.")
 
     @classmethod
     def add_args(cls, parser):
