+++ tests/test_project_legacy_api.py
@@ -13,6 +13,7 @@ from flow import label
 from flow import classlabel
 from flow import staticlabel
 from flow import init
+from flow.legacy_templating import JobScript
 from test_project import redirect_stdout, redirect_stderr, suspend_logging, \
     MockEnvironment, MockScheduler, StringIO
 
@@ -81,6 +82,16 @@ class LegacyMockProject(MockProject):
 
     def write_script_header(self, script, **kwargs):
         super(LegacyMockProject, self).write_script_header(script, **kwargs)
+
+
+def expect_deprecation_warnings(func):
+
+    def wrapper(self):
+        if six.PY2:
+            raise unittest.SkipTest("Not implemented for Python 2.7.")
+        with self.assertWarns(DeprecationWarning):
+            return func(self)
+    return wrapper
 
 
 class ProjectTest(unittest.TestCase):
@@ -189,6 +200,7 @@ class ProjectTest(unittest.TestCase):
         for job in project:
             self.assertIn('said_hello', list(project.labels(job)))
 
+    @expect_deprecation_warnings
     def test_single_submit(self):
         env = get_environment()
         env.scheduler_type.reset()
@@ -218,7 +230,7 @@ class ProjectTest(unittest.TestCase):
         self.assertEqual(len(list(sched.jobs())), 1)
         sched.reset()
 
-    @unittest.expectedFailure
+    @expect_deprecation_warnings
     def test_submit(self):
         env = get_environment()
         sched = env.scheduler_type()
@@ -231,7 +243,6 @@ class ProjectTest(unittest.TestCase):
         self.assertEqual(len(list(sched.jobs())), len(project))
         sched.reset()
 
-    @unittest.expectedFailure
     def test_submit_limited(self):
         env = get_environment()
         sched = env.scheduler_type()
@@ -245,7 +256,7 @@ class ProjectTest(unittest.TestCase):
                 project.submit(env, num=1)
                 self.assertEqual(len(list(sched.jobs())), 2)
 
-    @unittest.expectedFailure
+    @expect_deprecation_warnings
     def test_resubmit(self):
         env = get_environment()
         sched = env.scheduler_type()
@@ -278,7 +289,7 @@ class ProjectTest(unittest.TestCase):
                 project.submit(bundle_size=0)
                 self.assertEqual(len(list(sched.jobs())), 1)
 
-    @unittest.expectedFailure
+    @expect_deprecation_warnings
     def test_submit_status(self):
         env = get_environment()
         sched = env.scheduler_type()
@@ -309,17 +320,52 @@ class ProjectTest(unittest.TestCase):
                 fn_ = os.path.join(self._tmp_dir.name, fn)
                 self.assertTrue(os.path.isfile(fn_))
 
-    @unittest.skipIf(__name__ != '__main__', 'can only be tested if __main__')
-    def test_main(self):
-        project = self.mock_project()
-        with redirect_stderr(StringIO()):
-            with redirect_stdout(StringIO()):
-                with self.assertRaises(SystemExit):
-                    project.main()
+    @expect_deprecation_warnings
+    def test_JobScript(self):
+        env = get_environment(test=True)
+        sscript = env.script()
+        sscript_ = JobScript(env)
+        self.assertEqual(type(sscript), type(sscript_))
+        self.assertEqual(sscript._env, sscript_._env)
+        self.assertEqual(len(sscript.read()), 0)
+        sscript.writeline('test')
+        sscript.seek(0)
+        self.assertEqual(sscript.read(), 'test\n')
+        sscript = env.script()
+        sscript.write_cmd('test')
+        sscript.seek(0)
+        self.assertEqual(sscript.read(), 'test\n')
+        sscript = env.script()
+        sscript.write_cmd('test', bg=True)
+        sscript.seek(0)
+        self.assertTrue(sscript.read().endswith('&\n'))
+
+    @expect_deprecation_warnings
+    def test_write_test_submission_script(self):
+        env = get_environment(test=True)
+        sscript = env.script()
+        self.assertTrue(isinstance(sscript, JobScript))
+        sscript = env.script(a=0)
+        sscript.seek(0)
+        self.assertEqual(sscript.read(), '#TEST a=0\n')
+
+    @expect_deprecation_warnings
+    def test_submit_test_submission_script(self):
+        env = get_environment(test=True)
+        sscript = env.script(a=0)
+        sscript.seek(0)
+        tmp_out = StringIO()
+        with redirect_stdout(tmp_out):
+            env.submit(sscript, hold=True)
+        tmp_out.seek(0)
 
 
 class LegacyProjectTest(ProjectTest):
     project_class = LegacyMockProject
+
+    @expect_deprecation_warnings
+    def test_script(self):
+        super().test_script()
 
 
 if __name__ == '__main__':
